move and split pipeline clone / config
